+++ autotests/core/fetchjobtest.cpp
@@ -115,7 +115,7 @@ private Q_SLOTS:
 
         auto job = new TestFetchJob(scenarios.first().requestUrl);
         QVERIFY(execJob(job));
-        QCOMPARE(job->error(), scenarios.last().responseCode == 200 ? KGAPI2::NoError : scenarios.last().responseCode);
+        QCOMPARE(static_cast<int>(job->error()), scenarios.last().responseCode == 200 ? KGAPI2::NoError : scenarios.last().responseCode);
         QCOMPARE(job->response(), scenarios.last().responseData);
 
         QVERIFY(!FakeNetworkAccessManagerFactory::get()->hasScenario());
@@ -145,7 +145,7 @@ private Q_SLOTS:
         auto account = AccountPtr::create(QStringLiteral("MockAccount"), QStringLiteral("MockToken"));
         auto job = new TestFetchJob(account, scenarios.first().requestUrl);
         QVERIFY(execJob(job));
-        QCOMPARE(job->error(), scenarios.last().responseCode == 200 ? KGAPI2::NoError : scenarios.last().responseCode);
+        QCOMPARE(static_cast<int>(job->error()), scenarios.last().responseCode == 200 ? KGAPI2::NoError : scenarios.last().responseCode);
         QCOMPARE(job->response(), scenarios.last().responseData);
 
         QVERIFY(!FakeNetworkAccessManagerFactory::get()->hasScenario());
